backend/email-faq-agent/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic[email]==2.10.3
pydantic-settings==2.6.1
aiosmtplib==3.0.2
jinja2==3.1.4
premailer==3.10.0
aiokafka==0.11.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.0
